ui-charts: move mouse related hooks in common folder
