+++ front/ui/ui-charts/src/chronogram/components/ChronogramCanvas.tsx
@@ -6,12 +6,12 @@ import { OccupancyBlocksLayer } from './OccupancyBlocksLayer';
 import { TimeChartCanvasContext } from '../../common/context';
 import { getTimeToPixel, getPixelToTime } from '../../common/helpers/time';
 import { useCanvas } from '../../common/hooks/useCanvas';
+import { useMouseInteractions } from '../../common/hooks/useMouseInteractions';
+import { useMouseTracking } from '../../common/hooks/useMouseTracking';
 import { useSize } from '../../common/hooks/useSize';
 import { TimeCaptions } from '../../common/layers/TimeCaptions';
 import TimeGraduations from '../../common/layers/TimeGraduations';
 import type { PickingElement, Point } from '../../common/types';
-import { useMouseInteractions } from '../../spaceTimeChart/hooks/useMouseInteractions';
-import { useMouseTracking } from '../../spaceTimeChart/hooks/useMouseTracking'; //! bouger les hooks dans common
 import { DEFAULT_THEME } from '../../spaceTimeChart/lib/consts';
 import { MouseContext } from '../../spaceTimeChart/lib/context';
 import { validateTheme } from '../../spaceTimeChart/lib/theme';
