+++ front/ui/ui-charts/src/spaceTimeChart/components/SpaceTimeChart.tsx
@@ -6,12 +6,12 @@ import SpaceGraduations from './SpaceGraduations';
 import { TimeChartCanvasContext } from '../../common/context';
 import { getTimeToPixel, getPixelToTime } from '../../common/helpers/time';
 import { useCanvas } from '../../common/hooks/useCanvas';
+import { useMouseInteractions } from '../../common/hooks/useMouseInteractions';
+import { useMouseTracking } from '../../common/hooks/useMouseTracking';
 import { useSize } from '../../common/hooks/useSize';
 import { TimeCaptions } from '../../common/layers/TimeCaptions';
 import TimeGraduations from '../../common/layers/TimeGraduations';
 import type { PickingElement } from '../../common/types';
-import { useMouseInteractions } from '../hooks/useMouseInteractions';
-import { useMouseTracking } from '../hooks/useMouseTracking';
 import { DEFAULT_THEME } from '../lib/consts';
 import { MouseContext, SpaceTimeChartCanvasContext, SpaceTimeChartContext } from '../lib/context';
 import { validateTheme } from '../lib/theme';
